+++ src/dune_lang/action.ml
@@ -171,7 +171,7 @@ type t =
   | Diff of (String_with_vars.t, String_with_vars.t) Diff.t
   | No_infer of t
   | Pipe of Outputs.t * t list
-  | Cram of String_with_vars.t
+  | Cram of String_with_vars.t * Shell_spec.t
   | Patch of String_with_vars.t
   | Substitute of String_with_vars.t * String_with_vars.t
   | Withenv of String_with_vars.t Env_update.t list * t
@@ -352,8 +352,15 @@ let cstrs_dune_file t =
           Pipe (Outputs, ts) )
   ; ( "cram"
     , Syntax.since Stanza.syntax (2, 7)
-      >>> let+ script = sw in
-          Cram script )
+      >>> let+ script = sw
+          and+ shell_spec =
+            fields
+              (field
+                 "shell"
+                 (Syntax.since Stanza.syntax (3, 12) >>> Shell_spec.decode)
+                 ~default:Shell_spec.default)
+          in
+          Cram (script, shell_spec) )
   ]
 ;;
 
@@ -431,7 +438,9 @@ let rec encode =
   | No_infer r -> List [ atom "no-infer"; encode r ]
   | Pipe (outputs, l) ->
     List (atom (sprintf "pipe-%s" (Outputs.to_string outputs)) :: List.map l ~f:encode)
-  | Cram script -> List [ atom "cram"; sw script ]
+  | Cram (script, shell_spec) ->
+    let shell_spec = [ List (atom "shell" :: Shell_spec.encode shell_spec) ] in
+    List (atom "cram" :: sw script :: shell_spec)
   | Patch i -> List [ atom "patch"; sw i ]
   | Substitute (i, o) -> List [ atom "substitute"; sw i; sw o ]
   | Withenv (ops, t) ->
@@ -527,7 +536,7 @@ let rec map_string_with_vars t ~f =
   | Diff diff -> Diff (Diff.map diff ~path:f ~target:f)
   | No_infer t -> No_infer (map_string_with_vars t ~f)
   | Pipe (o, ts) -> Pipe (o, List.map ts ~f:(map_string_with_vars ~f))
-  | Cram sw -> Cram (f sw)
+  | Cram (script, shell_spec) -> Cram (f script, Shell_spec.map f shell_spec)
   | Patch i -> Patch (f i)
   | Substitute (i, o) -> Substitute (f i, f o)
   | Withenv (ops, t) ->