+++ src/dune_lang/dune_lang.ml
@@ -27,3 +27,4 @@ module Targets_spec = Targets_spec
 module Wrapped = Wrapped
 module Visibility = Visibility
 module Dep_conf = Dep_conf
+module Shell_spec = Shell_spec