+++ src/dune_rules/cram/cram_exec.mli
@@ -1,3 +1,9 @@
 open Import
 
-val action : Path.t -> Action.t
+type ('path, _) spec =
+  { script : 'path
+  ; shell_prog : 'path
+  ; shell_args : string list
+  }
+
+val action : (Path.t, Import.Path.Build.t) spec -> Action.t