+++ src/dune_rules/cram/cram_stanza.mli
@@ -9,6 +9,7 @@ type t =
   ; applies_to : applies_to
   ; alias : Alias.Name.t option
   ; deps : Dep_conf.t Bindings.t option
+  ; shell : Shell_spec.t
   ; enabled_if : Blang.t
   ; locks : Locks.t
   ; package : Package.t option
@@ -17,3 +18,4 @@ type t =
 type Stanza.t += T of t
 
 val decode : t Dune_lang.Decoder.t
+val system_shell_prog : ?loc:Loc.t -> context:Context.t -> [ `sh ] -> Action.Prog.t