+++ test/expect-tests/dune_lang/sexp_tests.ml
@@ -387,3 +387,57 @@ block
 comment|#
 |}
 ;;
+
+let print_shell_spec : Dune_lang.Shell_spec.t -> unit =
+  fun spec -> Dune_lang.pp Dune_lang.(List (Shell_spec.encode spec)) |> print
+;;
+
+let bash_path = Dune_lang.String_with_vars.make_text loc "/bin/bash"
+let zsh_path = Dune_lang.String_with_vars.make_text loc "/bin/zsh"
+let opt_x = Dune_lang.String_with_vars.make_text loc "-x"
+
+let%expect_test _ =
+  print_shell_spec System_shell;
+  [%expect {| (:system) |}]
+;;
+
+let%expect_test _ =
+  print_shell_spec (Custom_shell { prog = bash_path; args = [] });
+  [%expect {| (/bin/bash) |}]
+;;
+
+let%expect_test _ =
+  print_shell_spec (Custom_shell { prog = bash_path; args = [ opt_x ] });
+  [%expect {| (/bin/bash -x) |}]
+;;
+
+let%expect_test _ =
+  print_shell_spec (Custom_shell { prog = zsh_path; args = [] });
+  [%expect {| (/bin/zsh) |}]
+;;
+
+let print_action : Dune_lang.Action.t -> unit =
+  fun spec -> Dune_lang.pp (Dune_lang.Action.encode spec) |> print
+;;
+
+let cram_script = Dune_lang.String_with_vars.make_text loc "test.t"
+
+let%expect_test _ =
+  print_action (Cram (cram_script, System_shell));
+  [%expect {| (cram test.t (shell :system)) |}]
+;;
+
+let%expect_test _ =
+  print_action (Cram (cram_script, Custom_shell { prog = bash_path; args = [] }));
+  [%expect {| (cram test.t (shell /bin/bash)) |}]
+;;
+
+let%expect_test _ =
+  print_action (Cram (cram_script, Custom_shell { prog = bash_path; args = [ opt_x ] }));
+  [%expect {| (cram test.t (shell /bin/bash -x)) |}]
+;;
+
+let%expect_test _ =
+  print_action (Cram (cram_script, Custom_shell { prog = zsh_path; args = [] }));
+  [%expect {| (cram test.t (shell /bin/zsh)) |}]
+;;